Update fr.json
A better translation for French.

In French, we don't say : "Jean annoncer 'un sujet'"... In this case, this translation : "Jean annonce 'un sujet'" is better, but the best translation is : "Jean a annoncé 'un sujet'".

I know. French is a difficult language.

Thank you for your good work.

Best regards.